requirements.txt
# RecBole Benchmarking - Python Dependencies
#
# RECOMMENDED: Use split installation for faster setup
#   1. conda install --file requirements-conda.txt -c pytorch -c conda-forge
#   2. pip install -r requirements-pip.txt
#
# ALTERNATIVE: Full pip install (slower)
#   pip install -r requirements.txt

# Core ML Framework
torch>=2.0.0
numpy>=1.21.0
scipy>=1.7.0

# RecBole Recommendation Library
recbole>=1.1.1
ray[tune]>=2.0.0

# Data Processing
pandas>=1.3.0

# Configuration
pyyaml>=6.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Progress Tracking
tqdm>=4.62.0

# Utilities
scikit-learn>=1.0.0

# Optional: Experiment Tracking
# wandb>=0.12.0
# tensorboard>=2.8.0

# Optional: Jupyter Support
# jupyter>=1.0.0
# ipywidgets>=7.6.0
